+++ Vacuna/Block/Step1.php
@@ -1,0 +1,16 @@
+<?php
+namespace Farmacia\Vacuna\Block;
+class Step1 extends \Magento\Framework\View\Element\Template
+{
+	protected $_postFactory;
+	public function __construct(
+		\Magento\Framework\View\Element\Template\Context $context
+	)
+	{
+		parent::__construct($context);
+	}
+
+	
+       
+
+}
