--- Vacuna/Controller/Index/Display.php
@@ -1,20 +1,0 @@
-<?php
-namespace Farmacia\Vacuna\Controller\Index;
-
-class Display extends \Magento\Framework\App\Action\Action
-{
-	protected $_pageFactory;
-	public function __construct(
-		\Magento\Framework\App\Action\Context $context,
-		\Magento\Framework\View\Result\PageFactory $pageFactory)
-	{
-		$this->_pageFactory = $pageFactory;
-		return parent::__construct($context);
-	}
-
-	public function execute()
-	{
-	
-		return $this->_pageFactory->create();
-	}
-}
