+++ Vacuna/view/frontend/layout/vacuna_index_step1.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0"?>
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <referenceContainer name="content">
-        <block class="Farmacia\Vacuna\Block\Display" name="vacuna_display" template="Farmacia_Vacuna::sayhello.phtml" />
+        <block class="Farmacia\Vacuna\Block\Header" name="vacuna_header" template="Farmacia_Vacuna::header.phtml" />
+        <block class="Farmacia\Vacuna\Block\Step1" name="vacuna_step1" template="Farmacia_Vacuna::step1.phtml" />
     </referenceContainer>
 </page>
