+++ Vacuna/view/frontend/templates/header.phtml
@@ -1,0 +1,10 @@
+
+<?php
+
+/**
+ * @var \Farmacia\Vacuna\Block\Step1 $block
+ */
+
+echo $block->sayHello();
+
+?>
