--- Vacuna/view/frontend/templates/sayhello.phtml
@@ -1,13 +1,0 @@
-<?php
-
-/**
- * @var \Farmacia\Vacuna\Block\Display $block
- */
-
-echo $block->sayHello();
-
-?>
-
-
-</body>
-</html>
